Moving SendLogbookAction to logbook-ui-menu. Updated classpath.
